backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
openai
python-multipart
pydantic
python-dotenv